pyproject.toml
[project]
name = "nap-website-scripts"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "sentence-transformers",
    "umap-learn",
    "hdbscan",
    "Pillow",
    "numpy",
]
